bizfin-pro/requirements.txt
# BizFin Pro SEO Pipeline - Dependencies
# =====================================

# Core Framework
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
mysql-connector-python>=8.1.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web Scraping & Analysis
selenium>=4.15.0
webdriver-manager>=4.0.0
scrapy>=2.11.0
newspaper3k>=0.2.8

# Text Processing & NLP
nltk>=3.8.0
spacy>=3.7.0
textstat>=0.7.0
yake>=0.4.8

# SEO & Content Analysis
readability>=0.3.1
googletrans>=4.0.0
langdetect>=1.0.9

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.4.0
pyyaml>=6.0.0

# Logging & Monitoring
loguru>=0.7.0
sentry-sdk>=1.38.0

# HTTP & API
httpx>=0.25.0
aiohttp>=3.9.0
urllib3>=2.0.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Template Engine
jinja2>=3.1.0
markupsafe>=2.1.0

# JSON & Data Serialization
orjson>=3.9.0
ujson>=5.8.0

# Async Processing
asyncio>=3.4.3
aiofiles>=23.2.0
celery>=5.3.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Development Tools
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# WordPress Specific
python-wordpress-xmlrpc>=2.3
wordpress-api>=1.2.0

# SEO Tools
googlesearch-python>=1.2.0
serpapi>=0.1.0
ahrefs-api>=1.0.0

# Content Generation
openai>=1.3.0
anthropic>=0.7.0
transformers>=4.35.0
torch>=2.1.0

# Utilities
click>=8.1.0
rich>=13.6.0
tqdm>=4.66.0
schedule>=1.2.0
croniter>=1.4.0

# Security
cryptography>=41.0.0
bcrypt>=4.1.0

# Performance
psutil>=5.9.0
memory-profiler>=0.61.0

# File Processing
python-magic>=0.4.27
chardet>=5.2.0

# Email
smtplib2>=0.2.0
email-validator>=2.1.0

# Web Framework (for dashboard)
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0

# Additional SEO Tools
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Monitoring & Analytics
prometheus-client>=0.19.0
grafana-api>=1.0.0

# Backup & Recovery
boto3>=1.34.0
azure-storage-blob>=12.19.0

# Version Control
gitpython>=3.1.0

# System Information
platform>=1.0.8
distro>=1.8.0


